+++ frontend/src/components/user-preferences/ProfilePreferences.vue
@@ -47,7 +47,7 @@
 
 <script>
 import SavingAnimation from '@src/components/SavingAnimation.vue'
-import { onlyName } from '@src/lib/page_context'
+import { onlyName } from '@src/lib/page-context'
 
 export default {
   name: 'ProfilePreferences',
