+++ frontend/src/lib/printer-comm.js
