+++ frontend/src/lib/printer-local-comm.js
