+++ frontend/src/lib/viewing-throttle.js
