+++ frontend/src/lib/vue-swal.js
