+++ frontend/src/views/NewOctoPrintTunnelPage.vue
@@ -266,7 +266,7 @@ import axios from 'axios'
 import moment from 'moment'
 import urls from '@config/server-urls'
 import { normalizedPrinter } from '@src/lib/normalizers'
-import { user, settings } from '@src/lib/page_context'
+import { user, settings } from '@src/lib/page-context'
 import octopodLogo from '@static/img/octopod.webp'
 import octoappLogo from '@static/img/octoapp.webp'
 import polymerLogo from '@static/img/polymer.webp'
