+++ frontend/src/views/SharedPrinterPage.vue
@@ -38,7 +38,7 @@
 import split from 'lodash/split'
 import { normalizedPrinter } from '@src/lib/normalizers'
 import urls from '@config/server-urls'
-import PrinterComm from '@src/lib/printer_comm'
+import PrinterComm from '@src/lib/printer-comm'
 import WebRTCConnection from '@src/lib/webrtc'
 import StreamingBox from '@src/components/StreamingBox'
 import NavBar from '@src/components/NavBar.vue'
